requirements.txt
streamlit
langchain
langgraph
chromadb
sentence-transformers
transformers
torch
pymupdf
psycopg2-binary
